Enum gif::streaming_decoder::Decoded

#[non_exhaustive]
pub enum Decoded {
    Nothing,
    GlobalPalette(Box<[u8]>),
    BackgroundColor(u8),
    Repetitions(Repeat),
    HeaderEnd,
    BlockStart(Block),
    SubBlockFinished(AnyExtension),
    BlockFinished(AnyExtension),
    FrameMetadata(FrameDataType),
    BytesDecoded(NonZeroUsize),
    LzwDataCopied(usize),
    DataEnd,
}

Indicates whether a certain object has been decoded

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Nothing

Decoded nothing.

GlobalPalette(Box<[u8]>)

Global palette.

BackgroundColor(u8)

Index of the background color in the global palette.

Repetitions(Repeat)

Loop count is known

HeaderEnd

Palette and optional Application extension have been parsed, reached frame data.

BlockStart(Block)

The start of a block. BlockStart(Block::Trailer) is the very last decode event

SubBlockFinished(AnyExtension)

Decoded a sub-block. More sub-block are available.

Indicates the label of the extension which might be unknown. A label of 0 is used when the sub block does not belong to an extension.

Call last_ext() to get the data

BlockFinished(AnyExtension)

Decoded the last (or only) sub-block of a block.

Indicates the label of the extension which might be unknown. A label of 0 is used when the sub block does not belong to an extension.

Call last_ext() to get the data

FrameMetadata(FrameDataType)

Decoded all information of the next frame, except the image data.

The returned frame does not contain any owned image data.

Call current_frame_mut() to access the frame info.

BytesDecoded(NonZeroUsize)

Decoded some data of the current frame. Size is in bytes, always > 0

LzwDataCopied(usize)

Copied (or consumed and discarded) compressed data of the current frame. In bytes.

DataEnd

No more data available the current frame.