Struct gif::DecodeOptions

pub struct DecodeOptions { /* fields omitted */ }

Options for opening a GIF decoder.

Implementations

impl DecodeOptions

pub fn new() -> DecodeOptions

Creates a new decoder builder

pub fn set_color_output(&mut self, color: ColorOutput)

Configure how color data is decoded.

pub fn set_memory_limit(&mut self, limit: MemoryLimit)

Configure a memory limit for decoding.

pub fn check_frame_consistency(&mut self, check: bool)

Configure if frames must be within the screen descriptor.

The default is false.

When turned on, all frame descriptors being read must fit within the screen descriptor or otherwise an error is returned and the stream left in an unspecified state.

When turned off, frames may be arbitrarily larger or offset in relation to the screen. Many other decoder libraries handle this in highly divergent ways. This moves all checks to the caller, for example to emulate a specific style.

pub fn read_info<R: Read>(self, r: R) -> Result<Decoder<R>, DecodingError>

Reads the logical screen descriptor including the global color palette

Returns a Decoder. All decoder configuration has to be done beforehand.