Struct gif::Encoder

pub struct Encoder<W: Write> { /* fields omitted */ }

GIF encoder.

Methods

impl<W: Write> Encoder<W>

pub fn new(w: W, width: u16, height: u16, global_palette: &[u8]) -> Result<Self>

Creates a new encoder.

global_palette gives the global color palette in the format [r, g, b, ...], if no global palette shall be used an empty slice may be supplied.

pub fn write_global_palette(self, palette: &[u8]) -> Result<Self>

Writes the global color palette.

pub fn write_frame(&mut self, frame: &Frame) -> Result<()>

Writes a frame to the image.

Note: This function also writes a control extension if necessary.

pub fn write_extension(&mut self, extension: ExtensionData) -> Result<()>

Writes an extension to the image.

It is normally not necessary to call this method manually.

pub fn write_raw_extension(&mut self, func: u8, data: &[&[u8]]) -> Result<()>

Writes a raw extension to the image.

This method can be used to write an unsupported extesion to the file. func is the extension identifier (e.g. Extension::Application as u8). data are the extension payload blocks. If any contained slice has a lenght > 255 it is automatically divided into sub-blocks.