Struct gif::StreamingDecoder

pub struct StreamingDecoder {
    state: Option<State>,
    lzw_reader: Option<Decoder<LsbReader>>,
    skip_extensions: bool,
    version: &'static str,
    width: u16,
    height: u16,
    global_color_table: Vec<u8>,
    background_color: [u8; 4],
    ext: (u8, Vec<u8>, bool),
    current: Option<Frame<'static>>,
}

GIF decoder which supports streaming

Fields
state: Option<State> lzw_reader: Option<Decoder<LsbReader>> skip_extensions: bool version: &'static str width: u16 height: u16 global_color_table: Vec<u8> background_color: [u8; 4] ext: (u8, Vec<u8>, bool)

ext buffer

current: Option<Frame<'static>>

Frame data

Methods

impl StreamingDecoder

fn new() -> StreamingDecoder

Creates a new streaming decoder

fn update<'a>(&'a mut self,
buf: &[u8])
-> Result<(usize, Decoded<'a>), DecodingError>

Updates the internal state of the decoder.

Returns the number of bytes consumed from the input buffer and the last decoding result.

fn last_ext(&self) -> (u8, &[u8], bool)

Returns the data of the last extension that has been decoded.

fn current_frame_mut<'a>(&'a mut self) -> &'a mut Frame<'static>

Current frame info as a mutable ref.

fn current_frame<'a>(&'a self) -> &'a Frame<'static>

Current frame info as a ref.

fn width(&self) -> u16

Width of the image

fn height(&self) -> u16

Height of the image

fn next_state<'a>(&'a mut self,
buf: &[u8])
-> Result<(usize, Decoded<'a>), DecodingError>

fn read_control_extension(&mut self, b: u8) -> Result<State, DecodingError>

fn add_frame(&mut self)