Struct gif::streaming_decoder::StreamingDecoder

pub struct StreamingDecoder {

    state: State,
    lzw_reader: LzwReader,
    skip_frame_decoding: bool,
    check_frame_consistency: bool,
    allow_unknown_blocks: bool,
    memory_limit: MemoryLimit,
    version: Version,
    width: u16,
    height: u16,
    global_color_table: Vec<u8>,
    background_color: [u8; 4],
    ext: ExtensionData,
    current: Option<Frame<'static>>,
    header_end_reached: bool,

}

GIF decoder which emits low-level events for items in the GIF file

To just get GIF frames, use crate::Decoder instead.

Fields
state: Statelzw_reader: LzwReaderskip_frame_decoding: boolcheck_frame_consistency: boolallow_unknown_blocks: boolmemory_limit: MemoryLimitversion: Versionwidth: u16height: u16global_color_table: Vec<u8>background_color: [u8; 4]ext: ExtensionData

ext buffer

current: Option<Frame<'static>>

Frame data

header_end_reached: bool

Needs to emit HeaderEnd once

Implementations

impl StreamingDecoder

pub fn new() -> StreamingDecoder

Creates a new streaming decoder

pub(crate) fn with_options(options: &DecodeOptions) -> Self

pub fn update<'a>( &'a mut self, buf: &[u8], write_into: &mut OutputBuffer<'_>, ) -> Result<(usize, Decoded), DecodingError>

Updates the internal state of the decoder.

Returns the number of bytes consumed from the input buffer and the last decoding result.

pub fn last_ext(&self) -> (AnyExtension, &[u8], bool)

Returns the data of the last extension that has been decoded.

pub fn current_frame_mut(&mut self) -> &mut Frame<'static>

Current frame info as a mutable ref.

pub fn current_frame(&self) -> &Frame<'static>

Current frame info as a ref.

fn try_current_frame(&mut self) -> Result<&mut Frame<'static>, DecodingError>

Current frame info as a mutable ref.

pub fn width(&self) -> u16

Width of the image

pub fn height(&self) -> u16

Height of the image

pub fn version(&self) -> Version

The version number of the GIF standard used in this image.

We suppose a minimum of V87a compatibility. This value will be reported until we have read the version information in the magic header bytes.

fn next_state( &mut self, buf: &[u8], write_into: &mut OutputBuffer<'_>, ) -> Result<(usize, Decoded), DecodingError>

fn read_second_byte( &mut self, next: U16Value, value: u8, b: u8, ) -> Result<State, DecodingError>

fn read_control_extension(&mut self, b: u8) -> Result<State, DecodingError>

fn add_frame(&mut self)