Struct gif::Reader

pub struct Reader<R: Read> { /* fields omitted */ }

GIF decoder

Methods

impl<R> Reader<R> where R: Read

fn next_frame_info(&mut self) -> Result<Option<&Frame<'static>>, DecodingError>

Returns the next frame info

fn read_next_frame(&mut self) -> Result<Option<&Frame<'static>>, DecodingError>

Reads the next frame from the image.

Do not call Self::next_frame_info beforehand. Deinterlaces the result.

fn read_into_buffer(&mut self, buf: &mut [u8]) -> Result<(), DecodingError>

Reads the data of the current frame into a pre-allocated buffer.

Self::next_frame_info needs to be called beforehand. The length of buf must be at least Self::buffer_size. Deinterlaces the result.

fn fill_buffer(&mut self, buf: &mut [u8]) -> Result<bool, DecodingError>

Reads data of the current frame into a pre-allocated buffer until the buffer has been filled completely.

Self::next_frame_info needs to be called beforehand. Returns true if the supplied buffer could be filled completely. Should not be called after false had been returned.

fn buffer_size(&self) -> usize

Output buffer size

fn line_length(&self) -> usize

Line length of the current frame

fn palette(&self) -> Result<&[u8], DecodingError>

Returns the color palette relevant for the current (next) frame

fn global_palette(&self) -> Option<&[u8]>

The global color palette

fn width(&self) -> u16

Width of the image

fn height(&self) -> u16

Height of the image

fn bg_color(&self) -> Option<usize>

Index of the background color in the global palette