Struct gif::StreamingDecoder

pub struct StreamingDecoder { /* fields omitted */ }

GIF decoder which supports streaming

Methods

impl StreamingDecoder

fn new() -> StreamingDecoder

Creates a new streaming decoder

fn update<'a>(&'a mut self,
buf: &[u8])
-> Result<(usize, Decoded<'a>), DecodingError>

Updates the internal state of the decoder.

Returns the number of bytes consumed from the input buffer and the last decoding result.

fn last_ext(&self) -> (u8, &[u8], bool)

Returns the data of the last extension that has been decoded.

fn current_frame_mut<'a>(&'a mut self) -> &'a mut Frame<'static>

Current frame info as a mutable ref.

fn current_frame<'a>(&'a self) -> &'a Frame<'static>

Current frame info as a ref.

fn width(&self) -> u16

Width of the image

fn height(&self) -> u16

Height of the image